backend/requirements.txt
flask
firebase-admin
xrpl-py>=4.0.0
python-dotenv
cryptoconditions
